requirements.txt
openai==0.27.0
requests==2.31.0
imapclient==2.3.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pytest==7.4.0
black==23.7.0
isort==5.12.0
